Team_SIA_Birthstone_with_nested_functions.cpp; Birthstone 12.cpp
Hey, tried to get the nested functions to work but it just skips right over the functions and sends the user to the end of the program :/ hoping maybe you could look at it and see what the problem is. The "birthstone 12" file is the code I refactored. Fixed some of the verbage in there as well as added (ans == 'Y' || and == 'y'), we only had it accepting uppercase Y from the user, now it will take either.